% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploytools.R
\name{build_deploy_website}
\alias{build_deploy_website}
\title{build and deploy website}
\usage{
build_deploy_website(pkg = ".", psFrom = file.path(pkg, "vignettes",
  "index.Rmd"), psToDir = ".")
}
\arguments{
\item{pkg}{package directory where vignette sources are stored}

\item{psFrom}{directory where vignette source is stored}

\item{psToDir}{target directory to where vigette is deployed to}
}
\description{
First any existing output of vignettes in pkg/inst/doc is removed
using \code{devtools::clean_vignettes}. Then vignettes are built
using \code{devtools::build_vignettes}. In the last step, the result
is deployed to a given target directory using \code{deploy_website}.
}

