% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{cleanup_slidesdir}
\alias{cleanup_slidesdir}
\title{Clean up output files from slide compilation}
\usage{
cleanup_slidesdir(psSlidesDir = "vignettes", psFormatToKeep = c("rmd$",
  "rnw$", "cls$", "rpres$"), psDirsToKeep = c("ETH-BG", "odg", "png", "tex"))
}
\arguments{
\item{psSlidesDir}{root directory of document source files}

\item{psFormatToKeep}{file extensions of source files to be ignored by cleanup}

\item{psDirsToKeep}{explicit names of directories to be ignored by cleanup}
}
\description{
Cleaning up files and directories that are produced while compiling different
types of rmarkdown sources. If certain files should be ignored by the cleanup
function, then their extension can be specified as a component of the vector
passed as function parameter psFormatToKeep. If directories should be kept,
then their name can be specified as a component of the parameter vector
psDirsToKeep.
}
\details{
As a security feature, a question to the user is asked whether or not to delete
the output files.
}

