% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{cleanup_source_notes}
\alias{cleanup_source_notes}
\title{Clean up notes source files}
\usage{
cleanup_source_notes(psNotesPattern = "_Notes.Rmd$", psSubDir = "vignettes",
  psNotesSrcFile = NULL)
}
\arguments{
\item{psNotesPattern}{pattern of notes file to be deleted}

\item{psSubDir}{subdirectory where notes file is}

\item{psNotesSrcFile}{explicit name of notes file}
}
\description{
This function \code{cleanup_source_notes()} is to be used with caution,
because it will delete any Rmd source file of slide notes. It is created
for convenience in the testing phase with notes. Once slide notes with
meaningful content are created, it should not be used anymore.
}

