% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-refclass.R
\docType{class}
\name{CourseWebsiteRefClass-class}
\alias{CourseWebsiteRefClass}
\alias{CourseWebsiteRefClass-class}
\title{Reference Class CourseWebsiteRefClass implements a concrete representation of a website}
\description{
All attributes and methods of "GenericDocumentRefClass" are inherited. The main method
in this reference class is to initialize a new website
}
\section{Methods}{

\describe{
\item{\code{initWebSite(psTemplate, psPath)}}{Initialize a new website}
}}

