% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenttools.R
\name{create_course_document}
\alias{create_course_document}
\title{Create a course document}
\usage{
create_course_document(psDocuName, psPkgPath = ".",
  psRmdTemplate = "pdf_document", psTemplatePkg = "rcoursetools",
  psDocuSubdir = "vignettes", pbEdit = FALSE)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbEdit}{directly open newly created document}
}
\description{
\code{create_course_document} assumes that psPkgPath is a
directory that contains an R-package. By default the new
document is created in subdirectory "vignettes". If this
subdirectory does not exist, it is created. The document
is generated by the function \code{rmarkdown::draft} using
the template "pdf_document".
}
\details{
The basic functionality follows the function
\code{devtools::use_vignette}, except for the possibility
of specifying any given template from any package. The template
can be specified via argument "psRmdTemplate". This requires
in the package given by psTemplatePkg that a subdirectory with
the name of the template exists under "rmarkdown/templates".
More information about setting up custom templates is obtained
in the helpfile of rmarkdown::draft.
}

