% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenttools.R
\name{create_course_website}
\alias{create_course_website}
\title{Create a course website}
\usage{
create_course_website(psCourseName, psCourseDir = ".", psWsTemplate = NULL)
}
\arguments{
\item{psCourseName}{Name of the course}

\item{psCourseDir}{Course directory}

\item{psWsTemplate}{Website template}
}
\description{
Create a course website
}

