% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramtools.R
\name{create_diagram}
\alias{create_diagram}
\title{Create an empty diagram}
\usage{
create_diagram(psFormat = "odg")
}
\arguments{
\item{psFormat}{Format of diagram to be created}
}
\description{
\code{create_diagram} uses a templated for a given format
that is expected to be stored in inst/extdata/templates
}

