% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{create_notes}
\alias{create_notes}
\title{Create Notes document for a given deck of slides}
\usage{
create_notes(psSlideSrcFile, pbEditNotesFile = TRUE)
}
\arguments{
\item{psSlideSrcFile}{Rmd source file with slides}

\item{pbEditNotesFile}{Flag indicating whether to open the notes file for edit}
}
\description{
For a given Rmd source file containing a deck of
slides, a new Rmd files is created. The new file
contains all the slides from the source file with
an extra slides inserted between each existing
slide. The newly inserted slides all have the title
Notes. These new slides can be used for notes
used during the presentation. The file with the
notes has the same name as the slide source file
with the string "_Notes" added before the extension.
}

