% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenttools.R
\name{create_rmd_document}
\alias{create_rmd_document}
\title{Create a document from a template}
\usage{
create_rmd_document(psDocuName, psDocuPath = ".",
  psRmdTemplate = "pdf_document", psTemplatePkg = "rcoursetools",
  pbCreateDocuSubdir = TRUE, pbEdit = TRUE)
}
\arguments{
\item{psDocuName}{name of the new document without extension}

\item{psDocuPath}{path where document should be created}

\item{psRmdTemplate}{name of the template to be used}

\item{psTemplatePkg}{package from where the template should be taken}

\item{pbCreateDocuSubdir}{separate subdirectory for document psDocuName}

\item{pbEdit}{directly open newly created document}
}
\description{
The function \code{create_rmd_document} copies a template document
from any given package to the path that is specified using
the name that is given with parameter \code{psDocuName}.
}
\details{
The format of the created document is assumed to be Rmarkdown.
The extension .Rmd will be added to the given document name.
This function does not make any assumption about the environment
in which the new document is created. Hence this function can
be used outside of R-Packages or RStudio-projects.
}

