% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{create_slides}
\alias{create_slides}
\title{Create a deck of slides}
\usage{
create_slides(psSlidesName, psCourseDir = ".",
  psSlidesTemplate = "beamer_slides", psSlidesFormat = "Rmd")
}
\arguments{
\item{psSlidesName}{Name of Source-File for Slides}

\item{psCourseDir}{Course directory}

\item{psSlidesTemplate}{template for slides which should be used}

\item{psSlidesFormat}{extension of final slides file name}
}
\description{
This is a wrapper function to \code{rmarkdown::draft} that
creates a source document for a deck of beamer slides. The
template is taken from this packages template directory
}
\details{
The setup here is done analogeously to \code{devtools::use_vignette},
except that here, we are using different templates and we want to
create pdf_ouput and not html_vignette. The template to be used
can be specified as a parameter, by default "beamer_slides" is
used as a template.
}

