% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templatetools.R
\name{create_template}
\alias{create_template}
\title{Create skeleton directories and files for a new template}
\usage{
create_template(psTemplateName, psPkgDir = ".")
}
\arguments{
\item{psTemplateName}{name of the template to be created}

\item{psPkgDir}{Package directory}
}
\description{
The function \code{rmarkdown::draft()} uses a defined
directory structure under "rmarkdown/templates" to
instantiate new documents. This function \code{create_template()}
creates a new skeleton infrastructure for a new template.
}
\details{
In principle, we are doing meta-templating here. We use a
generic template and create a more specific template based
on that. The generic template is stored as template of
this package "rcoursetools". From this the template
infrastructure is copied to the target package from where
it can be used as ordinary template.
}

