% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcfperson-refclass.R
\docType{class}
\name{DcfAuthorRefClass-class}
\alias{DcfAuthorRefClass}
\alias{DcfAuthorRefClass-class}
\title{Reference class to represent an author in a description file. This refclass is derived from DcfPersonRefClass}
\description{
Reference class to represent an author in a description file. This refclass is derived from DcfPersonRefClass
}
\section{Methods}{

\describe{
\item{\code{addRole(psRole)}}{adding a role to current list of roles}

\item{\code{toDcfString()}}{writing author information to dcf-formatted string}
}}

