% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcfperson-refclass.R
\docType{class}
\name{DcfMaintainerRefClass-class}
\alias{DcfMaintainerRefClass}
\alias{DcfMaintainerRefClass-class}
\title{Reference Class to represent a maintainer in a dcf file, derived from DcfPersonRefClass}
\description{
Reference Class to represent a maintainer in a dcf file, derived from DcfPersonRefClass
}
\section{Methods}{

\describe{
\item{\code{toDcfString()}}{writing maintainer information to dcf-formatted string}
}}

