% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dcfperson-refclass.R
\docType{class}
\name{DcfPersonRefClass-class}
\alias{DcfPersonRefClass}
\alias{DcfPersonRefClass-class}
\title{Abstract reference class for a person in a description file}
\description{
Reference objects from class \code{DcfPersonRefClass} can be used
for authors or maintainers
}

