% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-refclass.R
\docType{class}
\name{DocumentSourceRefClass-class}
\alias{DocumentSourceRefClass}
\alias{DocumentSourceRefClass-class}
\title{Representation of document sources}
\description{
This reference class characterizes the sources for a given document.
Documents are either generated from LaTeX- or Markdown-sources. The
necessary infomration that characterize the source of a certain document
are contained in this reference class.
}

