% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-refclass.R
\docType{class}
\name{GenericDocumentRefClass-class}
\alias{GenericDocumentRefClass}
\alias{GenericDocumentRefClass-class}
\title{Common attributes of documents are grouped in this reference class}
\description{
Documents are either in latex or in markdown format. The common
characteristics of such documents are contained in this reference
class.
}

