% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_helper.R
\name{getPandocPath}
\alias{getPandocPath}
\title{Try to find path to pandoc}
\usage{
getPandocPath()
}
\description{
Especially on Mac Os X system paths are no longer propagated to applications,
hence we try to do some work in order to find a pandoc executable
}

