% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramtools.R
\name{open_odg_graphic}
\alias{open_odg_graphic}
\title{Open an existing odg graphics file}
\usage{
open_odg_graphic(psOdgGraphicName)
}
\arguments{
\item{psOdgGraphicName}{Name of the odg graphic file}
}
\description{
Existing graphic files in odg format can be opened using
\code{open_odg_graphic}. Depending on the OS platform
different commands are used to open the graphic file which
is specified by the parameter psOdgGraphicName
}

