% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_helper.R
\name{pandoc}
\alias{pandoc}
\title{run pandoc on an input document}
\usage{
pandoc(input, format, config = getOption("config.pandoc"), ext = NA,
  encoding = getOption("encoding"))
}
\description{
run pandoc on an input document
}

