% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler6class.R
\docType{data}
\name{R6DocuStatus}
\alias{R6DocuStatus}
\title{================================================================================== '#
Reference class representing document status information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6DocuStatus
}
\description{
Concrete class for document status table
}
\section{Fields}{

\describe{
\item{\code{sStatusFile}}{name of the status file}
}}
\keyword{datasets}

