% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler6class.R
\docType{data}
\name{R6GenericTable}
\alias{R6GenericTable}
\title{R6 class representing a generic table}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6GenericTable
}
\description{
Within the object of R6 class \code{RCGenericTable}
a table is represented by a header and a body. Tables can only be
modified by adding new rows. The two methods \code{to_knitr_kable}
and \code{to_pander_pandoc} write tables in markdown format using
functions \code{knitr::kable()} and \code{pander::pandoc.table()}
}
\details{
An object of R6 class \code{R6GenericTable} assumes
that a table simply consists of a header and a body. The body
is just a character vector where each component represents one
column header. The body is a list which each element contains the
data for one column. The R6 class components are represented
by the fields \code{sTableHeader} and \code{lTableBody}. The
initialisation method creates a table with an empty body. The table
header field has a getter and a setter methods. The method \code{addRow}
adds a new row at the end of the table.
}
\section{Fields}{

\describe{
\item{\code{sTableHeader}}{vector of table headers}

\item{\code{lTableBody}}{list of table body}
}}
\keyword{datasets}

