% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc-refclass.R
\docType{class}
\name{RCDesc-class}
\alias{RCDesc}
\alias{RCDesc-class}
\title{Reference class representing R package description objects}
\description{
The RCDesc reference class provides fields and methods that
can be used to read, to write and to extend DESCRIPTION files
of a package. The format of the DESCRIPTION files is explained
in the help files for \code{read.dcf}.
}
\section{Methods}{

\describe{
\item{\code{addToDcf()}}{Add DCF information to existing description}

\item{\code{readDcf()}}{Reading a description file}

\item{\code{writeDcf()}}{Write DCF information to a file}

\item{\code{writeToExDcf()}}{read a description file, add existing information and write back to desc file}
}}

