% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcoursetools-package.r
\docType{package}
\name{rcoursetools}
\alias{rcoursetools}
\alias{rcoursetools-package}
\title{rcoursetools - A Set Of Tools Useful For Teaching Courses.}
\description{
When teaching a course, one faces a standard set of tasks such as
\itemize{
  \item creating slides
  \item creating excercises
  \item creating homeworks
  \item grading homeworks
  \item answering student questions
  \item creating exams
}
Most of these tasks are repetitive and hence tools which support the
teachers in completing these tasks are believed to be valueable in
the sense that they take away some of the burden from the teachers
giving them more time to spend with students.
}
\details{
The package rcoursetools provides a set of tools which are meant to
help teachers in every step of the creation of a new course.
}

