% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_helper.R
\name{sans_ext}
\alias{sans_ext}
\title{remove file extension from x}
\usage{
sans_ext(x, compression = FALSE)
}
\arguments{
\item{x}{file name from which ending should be removed}

\item{compression}{flag indicating whether file x is compressed}
}
\value{
filename without extension
}
\description{
Extension of a file name is defined as the suffix that
occurs after the last dot (.). If argument \code{compression}
is set to true, then any endings that match "[.](gz|bz2|xz)$"
are removed first.
}

