% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_helper.R
\name{sub_ext}
\alias{sub_ext}
\title{substitute extension of file x with ext}
\usage{
sub_ext(x, ext)
}
\arguments{
\item{x}{vector with original file names}

\item{ext}{new extension to be added}
}
\value{
file names with new extensions
}
\description{
x is assumed to be a vector of characters which stand for
filenames. The extension of the file name is the suffix that
occurs after the last dot (.) in every component of x. The
substitution is done in two steps. First the existing extension
of x is removed using function \code{sans_ext}. Then the new
extension is added with a dot (.) as separator.
}

