% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{cleanup_slidesdir}
\alias{cleanup_slidesdir}
\title{Clean up output files from slide compilation}
\usage{
cleanup_slidesdir(psSlidesDir = "vignettes", psFormatToKeep = c("rmd$",
  "rnw$", "cls$", "rpres$"), psNamesToKeep = c("ETH-BG", "odg", "png", "tex"))
}
\arguments{
\item{psSlidesDir}{root directory of document source files}

\item{psFormatToKeep}{file extensions of source files to be ignored by cleanup}

\item{psNamesToKeep}{explicit names of files or directories to be ignored by cleanup}
}
\description{
Cleaning up files and directories that are produced while compiling different
types of rmarkdown sources
}

