% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coursetools.R
\name{create_course}
\alias{create_course}
\title{Create a new course}
\usage{
create_course(psCourseName, psCourseDir = ".")
}
\arguments{
\item{pscoursename}{name of the course}

\item{pscoursedir}{root directory where course is created}
}
\description{
The function \code{create_course} sets up the skeleton
infrastructure for a course
}

