% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_tools.R
\name{create_slides}
\alias{create_slides}
\title{Create a deck of slides}
\usage{
create_slides(psSlidesName, psCourseDir = ".", psSlidesSubdir = NULL)
}
\arguments{
\item{psSlidesName}{Name of Source-File for Slides}

\item{psCourseDir}{Course directory}
}
\description{
This is a wrapper function to \code{rmarkdown::draft} that
creates a source document for a deck of beamer slides. The
template is taken from this packages template directory
}

