% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidestools.R
\name{create_slides}
\alias{create_slides}
\title{Create a deck of slides}
\usage{
create_slides(psSlidesName, psCourseDir = ".", psSlidesSubdir = NULL,
  psSlidesTemplate = "beamer_slides", psSlidesFormat = "Rmd")
}
\arguments{
\item{psSlidesName}{name of slides source file}

\item{psCourseDir}{course directory}

\item{psSlidesSubdir}{subdirectory for current slides}

\item{psSlidesTemplate}{template for slides which should be used}

\item{psSlidesFormat}{extension of final slides file name}
}
\description{
This is a wrapper function to \code{rmarkdown::draft} that
creates a source document for a deck of beamer slides. The
template is taken from the template directory of this packages
}
\details{
The setup here is done analogeously to \code{devtools::use_vignette},
except that here, we are using different templates and we want to
create pdf_ouput and not html_vignette. The template to be used
can be specified as a parameter, by default "beamer_slides" is
used as a template. We explicitly allow and encourage source files of
slides to be organized in subdirectories of the vignettes directory,
knowing that this breaks vignette building by the devtools infrastructure.
There are several reasons for the subdirectories, first the function
\code{rmarkdown::draft} gives an error when a file or directory that
should be copied from the templated directory already exists in the
target directory where it should be copied to. Secondly, we want to
have the possibilities to use different documentclasses for different
slide documents. Thirdly, putting all slide sources of a complete
course in a single directory clutters the whole directory. Hence
organizing the material in subdirectories helps us to keep a better
overview over the whole course material.
}

