% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deploytools.R
\name{deploy_website}
\alias{deploy_website}
\title{Deploy a website from a source branch to a publication branch}
\usage{
deploy_website(psFrom = "vignettes/index.Rmd", psToDir = NULL)
}
\arguments{
\item{psFrom}{source file which is to be deployed, the source directory must exist, o/w, the function stops with an error}

\item{psToDir}{directory where the file should be deployed to, if it does not exist, the directory will be created}
}
\description{
\code{deploy_website} deploys a website that was created in some
branch directory into some other branch from where it is published
on the web via github pages. The need for this function comes from
the fact that github project pages are published from the gh-pages
branch and not from the master branch. But still we would like to
develop the website on the master branch which gives us the possibility
to test everyting without the need for immediate publication.
}

