% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramtools.R
\name{lGetDiagramDefaults}
\alias{lGetDiagramDefaults}
\title{Get a list with default settings related to diagrams}
\usage{
lGetDiagramDefaults()
}
\description{
Get a list with default settings related to diagrams
}

